% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background}
\alias{background}
\title{background}
\format{A data frame with 31995 rows and 1 variables:
\describe{
  \item{gene}{background gene list}
}}
\usage{
data(background)
background
}
\value{
A data frame
}
\description{
A data frame containing 31995 elements as background used for pathway
enrichment analysis:
}
\keyword{datasets}
