% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsNorm.R
\name{BetaNormScore}
\alias{BetaNormScore}
\title{Compute the beta-norm score}
\usage{
BetaNormScore(binomBayesDT, sdNorm, samplingRate)
}
\arguments{
\item{binomBayesDT}{Data table with binomial intervals}

\item{sdNorm}{A decimal indicating sigma of standard distribution normaliztion function}

\item{samplingRate}{Number of probability density function samples}
}
\value{
A \code{data.table} object containing score which is a 
result of a Bayesian binomial density probability function weighted
by normal distribution probability density function
}
\description{
Compute the beta-norm score
}
\examples{

conf = binom::binom.bayes(x=0:10,
                         n=10,
                         conf.level=0.95,
                         tol=.005,
                         type="central")
confDT = data.table::data.table(conf)
BetaNormScore(confDT, sdNorm=.2, samplingRate=100)

}
