% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{BScredIntervalCache}
\alias{BScredIntervalCache}
\title{Calculate credibility interval based on binomial bayes distribution with caching}
\usage{
BScredIntervalCache(bsData, cachedBinomialIntervals,
  methylCol = "methylCount", coverageCol = "coverage", confLevel = 0.95)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{cachedBinomialIntervals}{cachedBinomialIntervals}

\item{methylCol}{Name of column containing methylation count; defaults to "methylCount"}

\item{coverageCol}{Name of column containing coverage (i.e. number of reads); defaults to "coverage"}

\item{confLevel}{The level of confidence to be used in the confidence interval; default is 0.95}
}
\description{
Calculate credibility interval based on binomial bayes distribution with caching
}
