% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BSDTlist}
\alias{BSDTlist}
\title{A named \code{list} of 4 \code{data.table} objects, each of which has 10000 rows
containing 5 columns of methylation data for a given sample:
\itemize{
 \item{chr} {chromosome of methylation read}
 \item{start} {starting position for methylation read}
 \item{methylCount} {number of methylated reads}
 \item{coverage} {number of reads total}
 \item{sampleName} {name of the particular sample}
 }}
\format{A list object}
\usage{
data(BSDTlist)
}
\description{
A named \code{list} of 4 \code{data.table} objects, each of which has 10000 rows
containing 5 columns of methylation data for a given sample:
\itemize{
 \item{chr} {chromosome of methylation read}
 \item{start} {starting position for methylation read}
 \item{methylCount} {number of methylated reads}
 \item{coverage} {number of reads total}
 \item{sampleName} {name of the particular sample}
 }
}
\keyword{datasets}
