% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsNorm.R
\name{BSnormScoreCache}
\alias{BSnormScoreCache}
\title{Get credibility interval via binomial bayes distribution and caching}
\usage{
BSnormScoreCache(bsData, cachedBinomialIntervals,
  methylCol = "methylCount", coverageCol = "coverage",
  confLevel = 0.95, sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{cachedBinomialIntervals}{cachedBinomialIntervals}

\item{methylCol}{Name of column containing methylation count;
defaults to "methylCount"}

\item{coverageCol}{Name of column containing coverage (i.e. number of reads);
defaults to "coverage"}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}

\item{sdNorm}{A decimal indicating the standard deviation of the normal 
distribution density function used for score weighting (the CpGs 
with 50% methylation get always highest weight indicationg highest 
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score); 
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated 
density distributions sampled (how many points do we get in the 
interval 0-1 for each distribution); default is 100}
}
\value{
Bisulfite sequencing data as a \code{data.table} object with a column
    indicating score obtained by sum of Beta distribution density function (given
    by Byesian binomial shape parameters for given methylation) weighted by 
    predefined normal ditribution density function
}
\description{
Get credibility interval via binomial bayes distribution and caching
}
