% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{cacheBinomConfIntervals}
\alias{cacheBinomConfIntervals}
\title{Cache binomial confidence intervals}
\usage{
cacheBinomConfIntervals(maxHits, maxTotal, confLevel)
}
\arguments{
\item{maxHits}{maxHits}

\item{maxTotal}{maxTotal}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}
}
\description{
Cache binomial confidence intervals
}
