% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNorm.R
\name{calculateNormRPIM}
\alias{calculateNormRPIM}
\title{Helper function to get the relative proportion of flagged sites for a
single sample versus all other samples}
\usage{
calculateNormRPIM(sampleBaseline, sampleRelative, bsData,
  cacheDir = getOption("RESOURCES.RCACHE"), confLevel = 0.95,
  sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{sampleBaseline}{The sample (which should specify a name in the
bisulfite sequencing data) to use as the baseline
the proportion of sites for.}

\item{sampleRelative}{The sample to compare relatively to the baseline sample for
the proportion of sites.}

\item{bsData}{Bisulfite sequencing data for multiple samples; a BSDT
(bisulfite data.table) that has been split with splitDataTable
(so, a list of BSDTs); one corresponds to each sample to test.}

\item{cacheDir}{If using caching, this argument specifies the directory to
use for storing the cache;
defaults to global option for \code{RESOURCES.RACHE},
if no such option has been specified you must provide one}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}

\item{sdNorm}{A decimal indicating the standard deviation of the normal 
distribution density function used for score weighting (the CpGs 
with 50% methylation get always highest weight indicationg highest 
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score); 
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated 
density distributions sampled (how many points do we get in the 
interval 0-1 for each distribution); default is 100}
}
\value{
A vector of the same length as the number of samples being
analyzed; each element in the vector represents the the normalized proportion of
intermediate methylation relative to the other samples for a single sample
}
\description{
Helper function to get the relative proportion of flagged sites for a
single sample versus all other samples
}
