% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRPIM.R
\name{calculateRPIM}
\alias{calculateRPIM}
\title{Get the relative proportion of flagged sites. This is a general version
of a method to get the RPIM (Relative Proportion of Intermediate Methylation)
Given a DT with bisulfite reads, and a flag column,
and then a huge table with this data for lots of samples, this will calculate
the relative proportion of CGs that are flagged, for pairwise comparisons
between all samples, subsetted to the CGs present in both samples.}
\usage{
calculateRPIM(sampleName, BSDTsplit)
}
\arguments{
\item{sampleName}{The sample (which should specify a name in BSDTlist) to return
the proportion of sites for.}

\item{BSDTsplit}{A BSDT (bisulfite data.table) that has been split with
splitDataTable (so, a list of BSDTs); one corresponds to each sample to test.}
}
\description{
Get the relative proportion of flagged sites. This is a general version
of a method to get the RPIM (Relative Proportion of Intermediate Methylation)
Given a DT with bisulfite reads, and a flag column,
and then a huge table with this data for lots of samples, this will calculate
the relative proportion of CGs that are flagged, for pairwise comparisons
between all samples, subsetted to the CGs present in both samples.
}
