% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datTraits}
\alias{datTraits}
\alias{DEG}
\alias{DEH}
\alias{diffhetmatrix}
\alias{moduledm}
\alias{modulesil}
\alias{promoter}
\alias{sharedmatrix}
\title{datTraits}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{OS}{overall survival time}
  \item{EFS}{survival time}
  \item{age}{ages of patients}
}}
\usage{
data(datTraits)
datTraits

data(DEG)
DEG

data(DEH)
DEH

data(diffhetmatrix)
diffhetmatrix

data(moduledm)
moduledm

data(modulesil)
modulesil

data(promoter)
promoter

data(sharedmatrix)
sharedmatrix
}
\value{
A data frame

A data frame

A data frame

A data frame

A data frame

A data frame

A large GRanges object

A data frame
}
\description{
clinical traits containing OS,EFS,age

differentially expressed genes compared CEBPA-sil vs.normal

DEH loci

differentially heterogeneity matrix

module information for CEBPA-dm mutation samples

module information for CEBPA-sil mutation samples

promoter region annotation file

Epipolymorphism for 6 samples on DEH loci
}
\keyword{datasets}
