% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datTraits}
\alias{datTraits}
\title{datTraits}
\format{A data frame with 6 rows and 3 variables:
\describe{
  \item{OS}{overall survival time}
  \item{EFS}{survival time}
  \item{age}{ages of patients}
}}
\usage{
data(datTraits)
datTraits
}
\value{
A data frame
}
\description{
clinical traits containing OS,EFS,age
}
\keyword{datasets}
