% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEG}
\alias{DEG}
\title{DEG}
\format{A data frame with 13685 rows and 1 variables:
\describe{
  \item{refseq}{refseq ID of DEG}
}}
\usage{
data(DEG)
DEG
}
\value{
A data frame
}
\description{
differentially expressed genes compared CEBPA-sil vs.normal
}
\keyword{datasets}
