% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DEH}
\alias{DEH}
\title{DEH}
\format{A data frame with 479 rows and 3 variables:
\describe{
  \item{chromosome}{chromosome of DEH loci}
  \item{loci}{location of DEH loci}
  \item{strand}{strand of DEH loci}
}}
\usage{
data(DEH)
DEH
}
\value{
A data frame
}
\description{
DEH loci
}
\keyword{datasets}
