% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffHet.R
\name{diffHet}
\alias{diffHet}
\title{Calculate Differential Heterogeneity}
\usage{
diffHet(compare.matrix, value, group1, group2, subtype, het.dif.cutoff = 0.2,
  permutations = 1000, p.adjust.method = "fdr", cores = 5)
}
\arguments{
\item{compare.matrix}{The comparison matrix generated from
the compMatrix() function}

\item{value}{The value to be used in calculations. Possible
values are 'read', 'pdr', 'meth', 'epipoly', and 'shannon'}

\item{group1}{The first subtype group to be compared}

\item{group2}{The second subtype group to be compared}

\item{subtype}{A dataframe containing the subtype information
for the samples in the comparison matrix. The row names should
be the names of the samples and there should be one column
containing the subtype information for each sample.}

\item{het.dif.cutoff}{A number representing the cutoff
for the heterogeneity difference. If the heterogeneity difference is greater
than the cutoff value, than the p-value and adjusted p-value are
calculated for the loci. If the heterogeneity difference is less than
the cutoff value, than the p-value and adjusted p-value are set
to NA. (default: 0.20)}

\item{permutations}{The number of permutations for the
EntropyExplorer function. Value must be set to 'shannon'.
(default: 1000)}

\item{p.adjust.method}{The method to be used as a parameter in
p.adjust() function. Possible methods are 'holm', 'hochberg',
'hommel', 'bonferroni', 'BH', 'BY', 'fdr', and 'none'.(default: 'fdr')}

\item{cores}{The number of cores to be used for parallel execution.
Not available for 'shannon' values. (default: 5)}
}
\value{
A dataframe containing chromosome number, loci, mean of
group1, mean of group2, heterogeneity difference, and the p-value and
adjusted p-value for the loci with a heterogeneity difference greater
than the cutoff
}
\description{
From a user-inputted value and two subtype groups,
calculates the mean values for both subtypes at each
loci. The heterogeneity difference is calculated and the p-values
and adjusted p-values are calculated if the heterogeneity
difference is greater than a given cutoff
}
