% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diffhetmatrix}
\alias{diffhetmatrix}
\title{diffhetmatrix}
\format{A data frame with 30 rows and 8 variables:
\describe{
  \item{chromosome}{chromosome of loci}
  \item{loci}{location of loci}
  \item{CEBPA_sil.mean}{mean value of loci in CEBPA-sil patients}
  \item{Normal.mean}{mean value of loci in normal patients}
  \item{het.dif}{CEBPA_sil.mean-Normal.mean}
  \item{p.value}{p value}
  \item{p.adjust}{adjusted p value using multiple testing correction}
  \item{type}{the type of epigenetic heterogeneity:pdr,epipoly,shannon}
}}
\usage{
data(diffhetmatrix)
diffhetmatrix
}
\value{
A data frame
}
\description{
differentially heterogeneity matrix
}
\keyword{datasets}
