% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipathway.R
\name{epipathway}
\alias{epipathway}
\title{pathway annotation}
\usage{
epipathway(gene.list, cutoff = 0.05, prefix = NA, pdf.height = 10,
  pdf.width = 10)
}
\arguments{
\item{gene.list}{a data frame generated from network.construct() function.
The first column is gene entrez ID, the second column is module lable,
the third column is module color.}

\item{cutoff}{Cutoff value of qvalue for pathway enrichment}

\item{prefix}{a prefix for PDF file name}

\item{pdf.height}{An integer representing the height (in inches)
of the outputted boxplot pdf file (default: 10)}

\item{pdf.width}{An integer representing the width (in inches) of
the outputted boxplot pdf file (default: 10)}
}
\value{
a data frame containing pathways that are significantly enriched
by genes from one module
}
\description{
pathway identification significantly enriched by genes
in one module.
}
\examples{
genelist=data.frame(ENTREZID=c("2902","2905","3360","286223","59338",
"344018","5144","55001","7410","730051","55743","6804","200634","2802",
"2260","651","2104","23432","10505","23194","9855","7101",
"389136","124857","1829","3164","3754","8614","9469","3217","9578",
"10516","10630"),label=rep(18,33),color=rep("lightgreen",33),
stringsAsFactors = FALSE)
pathway = epihet::epipathway(genelist,cutoff = 0.05,
                             prefix="CEBPA_sil",pdf.height = 10,
                             pdf.width = 10)
}
