% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epiTSNE.R
\name{epiTSNE}
\alias{epiTSNE}
\title{Make TSNE Plot from Comparison Matrix}
\usage{
epiTSNE(compare.matrix, value, type, points.colors = NULL, theta = 0.5,
  curTheme = NULL, perplexity = 5, max_iter = 1000, pdf.height = 10,
  pdf.width = 10, sve = FALSE)
}
\arguments{
\item{compare.matrix}{The comparison matrix generated from
the compMatrix() function}

\item{value}{The value to be graphed in the PCA plot}

\item{type}{A dataframe containing the type information
for the samples in the comparison matrix. The row names should
be the names of the samples and there should be one column
containing the type information for each sample.}

\item{points.colors}{A vector of colors to be used as the color
of the individual points for each sample. One color is used per
subtype. (default: NULL)}

\item{theta}{A decimal representing the theta parameter for
the Rtsne() function. Represents the speed/accuracy trade-off
(0.0 is exact TSNE) (default: 0.5)}

\item{curTheme}{the theme of ggplot2 to control control the appearance of
all non-data components of the plot}

\item{perplexity}{An integer representing the perplexity
parameter for the Rtsne() function (default: 30)}

\item{max_iter}{An integer representing the max_iter parameter
for the Rtsne() function. Represents the number of iterations
(default: 1000)}

\item{pdf.height}{An integer representing the height (in inches)
of the outputted TSNE plot pdf file (default: 10)}

\item{pdf.width}{An integer representing the width (in inches) of
the outputted TSNE plot pdf file (default: 10)}

\item{sve}{A boolean to save the plot (default: FALSE)}
}
\value{
A T-SNE plot
}
\description{
From a user-inputted value, creates a TSNE plot from the
sample data and colors each point by the subtype information
provided.
}
\examples{
comp.Matrix=data.frame(
p1=c(0.6,0.3,0.5,0.5,0.5,0.6,0.45,0.57,0.45,0.63,0.58,0.67,0.5,0.42,0.67),
p2=c(0.62,0.63,0.55,0.75,0.84,0.58,1,0.33,1,0.97,0.57,0.68,0.73,0.72,0.82),
p3=c(0.72,0.53,0.62,0.69,0.37,0.85,1,0.63,0.87,0.87,0.82,0.81,0.79,
0.62,0.68),
N1=c(0.15,0.24,0.15,0.26,0.34,0.32,0.23,0.14,0.26,0.32,0.12,0.16,0.31,
0.24,0.32),
N2=c(0.32,0.26,0.16,0.36,0.25,0.37,0.12,0.16,0.41,0.47,0.13,0.52,0.42,
0.41,0.23),
N3=c(0.21,0.16,0.32,0.16,0.36,0.27,0.24,0.26,0.11,0.27,0.39,0.5,0.4,
0.31,0.33),
type=rep(c("pdr","epipoly","shannon"),c(5,5,5)),
location=rep(c("chr22-327:350:361:364","chr22-755:761:771:773",
"chr22-761:771:773:781","chr22-821:837:844:849","chr22-838:845:850:858"),
3),stringsAsFactors =FALSE )

subtype = data.frame(Type= c(rep('CEBPA_sil', 3), rep('Normal', 3)),
row.names = colnames(comp.Matrix)[1:6],stringsAsFactors = FALSE)

epihet::epiTSNE(compare.matrix = comp.Matrix, value = 'epipoly',
type = subtype, points.colors = NULL, theta = 0.5,
perplexity = 1, max_iter = 1000, pdf.height = 10,
pdf.width = 10, sve = TRUE)
}
