% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{getRPIM}
\alias{getRPIM}
\title{Get the relative proportion of flagged sites.}
\usage{
getRPIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"))
}
\arguments{
\item{bsData}{Bisulfite sequencing data for multiple samples; a BSDT (bisulfite data.table) that has been split with splitDataTable (so, a list of BSDTs); one corresponds to each sample to test.}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}
}
\description{
Get the relative proportion of flagged sites.
}
