% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHist.R
\name{histPIM}
\alias{histPIM}
\title{Given bisulfite sequencing data, calculates the proportion
of intermediate methylation sites from methylation percenatage
histogram and selected weighting function}
\usage{
histPIM(bsData, sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{bsData}{Bisulfite sequencing data;}

\item{sdNorm}{A decimal indicating the standard deviation of the normal
distribution density function used for score weighting (the CpGs
with 50% methylation get always highest weight indicationg highest
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score);
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated
density distributions sampled (how many points do we get in the
interval 0-1 for each distribution); default is 100}
}
\description{
Given bisulfite sequencing data, calculates the proportion
of intermediate methylation sites from methylation percenatage
histogram and selected weighting function
}
\examples{

data("exampleBSDT", package="epihet")


histPIM(exampleBSDT)
histPIM(exampleBSDT, sdNorm=.2, samplingRate=200)

}
