% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHist.R
\name{histRPIM}
\alias{histRPIM}
\title{Calculate the relative proportion of intermediate methylation (RPIM) score
from methylation histogram and weighting function}
\usage{
histRPIM(bsData, sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{bsData}{Bisulfite sequencing data for multiple samples. A list of BSDT
(bisulfite data.table), one corresponds to each sample to test. This may
also be a BSseq object}

\item{sdNorm}{A decimal indicating the standard deviation of the normal
distribution density function used for score weighting (the CpGs
with 50% methylation get always highest weight indicationg highest
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score);
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated
density distributions sampled (how many points do we get in the
interval 0-1 for each distribution); default is 100}
}
\value{
A named vector of the same length as the number of samples being
    analyzed; each element in the vector represents the the average
    proportion of intermediate methylation relative to the other samples; all
    samples are represented and elements named accordingly
}
\description{
Calculate the relative proportion of intermediate methylation (RPIM) score
from methylation histogram and weighting function
}
\examples{

data("BSDTlist", package="epihet")

histRPIM(BSDTlist)
histRPIM(BSDTlist, sdNorm=.2)

}
