% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesim.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard score calculation between modules from two subtypes.}
\usage{
jaccard(module.subtype1, module.subtype2)
}
\arguments{
\item{module.subtype1}{a data frame generated from the
epinetwork() function. The module information of subtype1,
the first column is module nodes, the second column is module label,
the third column is module color.}

\item{module.subtype2}{a data frame generated from
the epinetwork() function. The module information of subtype1,
the first column is module nodes, the second column is module label,
the third column is module color.}
}
\value{
A matrix containing Jaccard scores.
}
\description{
Jaccard score calculation based on the common genes in two
modules from two subtypes.
}
\examples{
data(modulesil)
data(moduledm)
jaccard.matrix = jaccard(modulesil, moduledm)
}
