% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGR.R
\name{makeGR}
\alias{makeGR}
\title{Make List of GenomicRanges Object}
\usage{
makeGR(files, ids, cores = 5, sve = FALSE)
}
\arguments{
\item{files}{A vector of input files containing methclone output files,
the suffix of files should be methClone_out.gz}

\item{ids}{A vector of sample ids for the files}

\item{cores}{The number of cores to be used for parallel execution
(default: 5)}

\item{sve}{A boolean to save the GenomicRanges object (default: FALSE)}
}
\value{
A list, each element is a data frame of GenomicRanges objects 
containing pdr, epipolymorphism, and Shannon entropy values for each 
input file. Saves as an epi.gr.rda extension
}
\description{
Creates a GenomicRanges object for each methclone output file
}
\examples{
path = system.file('extdata', package = 'epihet')
files = dir(path = path, pattern = 'methClone_out.gz',
            recursive = TRUE, full.names = TRUE)
ids = basename(dirname(files))
GR.List = epihet::makeGR(files = files, ids = ids,
cores = 1, sve = FALSE)
}
