% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methClone.R
\name{methClone}
\alias{methClone}
\title{A wrapper function for methclone package}
\usage{
methClone(file1, file2, outfile, sampleid, distance = 72, coverage = 60,
  methdiff = 0)
}
\arguments{
\item{file1}{File name for the first input of methclone package}

\item{file2}{File name for the second input of methclone package}

\item{outfile}{Output file name where methclone writes its output.
This is a gzipped file}

\item{sampleid}{Name of the sample provided to methClone package}

\item{distance}{Maximum distance between first and forth methylation sites
to be considered in a single pattern}

\item{coverage}{Minimum coverage at a signle methylation site to
be considered}

\item{methdiff}{Threshold for difference in methylation between the two
different samples}
}
\value{
Does not return any output, it writes methcone output into
the provided outfile name
}
\description{
Runs methclone package on the provided two BAM files
file
}
