% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moduleanno.R
\name{moduleanno}
\alias{moduleanno}
\title{module annotation}
\usage{
moduleanno(DEG, background, module.gene, cutoff = 0.05,
  adjust.method = "fdr", prefix = NA, pdf.height = 10, pdf.width = 10,
  sve = FALSE)
}
\arguments{
\item{DEG}{a character vector containing up/down regulated genes}

\item{background}{a charactor vector containing all genes as background
in hypergeometric test}

\item{module.gene}{a data frame containing genes with genome region
containing DEH loci from one module, generated from epinetwork() function.
The first column is gene entrez ID, the second column is module lable,
the third column is module color.}

\item{cutoff}{Cutoff value of qvalue for gene enrichment}

\item{adjust.method}{one of 'holm', 'hochberg', 'hommel', 'bonferroni',
'BH', 'BY', 'fdr', 'none'}

\item{prefix}{a prefix for PDF file name}

\item{pdf.height}{An integer representing the height (in inches)
of the outputted boxplot pdf file (default: 10)}

\item{pdf.width}{An integer representing the width (in inches) of
the outputted boxplot pdf file (default: 10)}

\item{sve}{A boolean to save the plot (default: FALSE)}
}
\value{
a data frame showing modules that were enriched
by DEGs and module size, p value and q value
}
\description{
annotate modules using differentially expressed genes.
}
\examples{
data(DEG,package = "epihet")
data(background,package = "epihet")
module=data.frame(gene=c("NM_000014","NM_000015","NM_000017","NM_000019",
"NM_052960","NR_138250","NM_000037","NM_000038","NM_000039","NM_000044",
"NM_000046","NM_015074","NM_183416","NM_004421","NM_001330311",
"NM_001145210","NM_000097","NM_000103","NM_000104",
"NM_000079","NM_000083","NM_000086","NM_000087","NM_000092","NM_000094",
"NM_000095","NM_006474"),label=rep(c(1,2),c(12,15)),
color=rep(c("purple","brown"),c(12,15)),
stringsAsFactors = FALSE)
module.annotation=epihet::moduleanno(DEG$refseq,background$gene,
                                     module.gene=module,
                                     cutoff=0.05,adjust.method = "fdr",
                                     prefix='epipoly',pdf.height = 10,
                                     pdf.width = 10, sve = TRUE)
}
