% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{moduledm}
\alias{moduledm}
\title{moduledm}
\format{A data frame with 140 rows and 3 variables:
\describe{
  \item{gene}{gene of module CEBPA_dm}
  \item{label}{label of module CEBPA_dm}
  \item{color}{color of module CEBPA_dm}
}}
\usage{
data(moduledm)
moduledm
}
\value{
A data frame
}
\description{
module information for CEBPA-dm mutation samples
}
\keyword{datasets}
