% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modulesil}
\alias{modulesil}
\title{modulesil}
\format{A data frame with 501 rows and 3 variables:
\describe{
  \item{gene}{gene of module CEBPA_sil}
  \item{label}{label of module CEBPA_sil}
  \item{color}{color of module CEBPA_sil}
}}
\usage{
data(modulesil)
modulesil
}
\value{
A data frame
}
\description{
module information for CEBPA-sil mutation samples
}
\keyword{datasets}
