% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modulesim.R
\name{modulesim}
\alias{modulesim}
\title{module comparison between two subtypes.}
\usage{
modulesim(module.subtype1, module.subtype2, pdf.height = 10, pdf.width = 10,
  sve = FALSE)
}
\arguments{
\item{module.subtype1}{a data frame generated from the network.construct()
function the module information of subtype1,the first column is module
nodes,the second column is module label, the third column is module color.}

\item{module.subtype2}{a data frame generated from the network.construct()
function. The module information of subtype1, the first column is
module nodes, the second column is module label, the third column is
module color.}

\item{pdf.height}{An integer representing the height (in inches)
of the outputted boxplot pdf file (default: 10)}

\item{pdf.width}{An integer representing the width (in inches) of
the outputted boxplot pdf file (default: 10)}

\item{sve}{A boolean to save the plot (default: FALSE)}
}
\value{
a matrix containing Jaccard scores
}
\description{
Compare any two modules from two subytpes based
on genes shared by the modules.
}
\examples{
data(modulesil,package = "epihet")
data(moduledm,package = "epihet")
sim.score=epihet::modulesim(module.subtype1=modulesil,
                            module.subtype2=moduledm,
                            pdf.height = 10,pdf.width = 10,
                            sve = TRUE)
}
