% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervalsNorm.R
\name{normCacheBinomScore}
\alias{normCacheBinomScore}
\title{Cache normalized score}
\usage{
normCacheBinomScore(maxHits, maxTotal, confLevel, sdNorm,
  samplingRate = 100)
}
\arguments{
\item{maxHits}{Maximum methylation count}

\item{maxTotal}{Maximum coverage}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}

\item{sdNorm}{A decimal indicating the standard deviation of the normal 
distribution density function used for score weighting (the CpGs 
with 50% methylation get always highest weight indicationg highest 
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score); 
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated 
density distributions sampled (how many points do we get in the 
interval 0-1 for each distribution); default is 100}
}
\value{
A \code{data.table} object containing a column for
score obtained by sum of weighted Bayesian binomial density function 
by normal distribution density function for maximum methylation count 
and coverage; this serves as a cache that can replace the need to 
perform the computationally expensive probability estimation
}
\description{
Cache normalized score
}
\examples{

normCacheBinomScore(100,100, confLevel=.95, sdNorm=.2)

}
