% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNorm.R
\name{normRPIM}
\alias{normRPIM}
\title{Calculate the relative proportion of intermediate methylation (RPIM) score.}
\usage{
normRPIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"),
  confLevel = 0.95, sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{bsData}{Bisulfite sequencing data for multiple samples. A list of BSDT
(bisulfite data.table), one corresponds to each sample to test. This may
also be a BSseq object}

\item{cacheDir}{If using caching, this argument specifies the directory to
use for storing the cache; defaults to global option for
\code{RESOURCES.RACHE}. If no such option has been specified you must
provide one}

\item{confLevel}{A decimal indicating the level of confidence to be used
while creating cached the binomial bayes credibility interval; default is
.95 for 95 percent confidence}

\item{sdNorm}{A decimal indicating the standard deviation of the normal
distribution density function used for score weighting (the CpGs
with 50% methylation get always highest weight indicationg highest
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score);
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated
density distributions sampled (how many points do we get in the
interval 0-1 for each distribution); default is 100}
}
\value{
A named vector of the same length as the number of samples being
    analyzed; each element in the vector represents the the average
    proportion of intermediate methylation relative to the other samples; all
    samples are represented and elements named accordingly
}
\description{
Calculate the relative proportion of intermediate methylation (RPIM) score.
}
\examples{

data("BSDTlist", package="epihet")

if (require(simpleCache)) {
normRPIM(BSDTlist, cacheDir="~")

simpleCache::setSharedCacheDir("cache")
normRPIM(BSDTlist)
normRPIM(BSDTlist, sdNorm=.2)
}
}
