% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{PIM}
\alias{PIM}
\title{Given bisulfite sequencing data, calculates the proportion
of intermediate methylation sites.}
\usage{
PIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"), imLower = 0.25,
  imUpper = 0.75)
}
\arguments{
\item{bsData}{Bisulfite sequencing data;}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}

\item{imLower}{The lower boundary for intermediate methylation (IM); if a site is entirely below this threshold (or if any part of a its binomial credibilty interval overlaps this boundary) it is not considered IM; defaults to .25}

\item{imUpper}{The upper boundary for intermediate methylation (IM); if a site is entirely above this threshold (or if any part of a its binomial credibilty interval overlaps this boundary) it is not considered IM; defaults to .75}
}
\description{
Given bisulfite sequencing data, calculates the proportion
of intermediate methylation sites.
}
