% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateHist.R
\name{prepHistRPIM}
\alias{prepHistRPIM}
\title{Helper function for the relative PIM (RPIM) score computed from
methylation histogram - takes coverage and methylation counts and
returns PIM score for these.
The function is used in calculateHistRPIM function.}
\usage{
prepHistRPIM(coverage, methylCount, sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{coverage}{Vector containing information about CpG coverage}

\item{methylCount}{Vector containing infomation about methylation counts
at CpG}

\item{sdNorm}{A decimal indicating the standard deviation of the normal
distribution density function used for score weighting (the CpGs
with 50% methylation get always highest weight indicationg highest
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score);
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated
density distributions sampled (how many points do we get in the
interval 0-1 for each distribution); default is 100}
}
\value{
A PIM score based on methylation histogram and weighting function
}
\description{
Helper function for the relative PIM (RPIM) score computed from
methylation histogram - takes coverage and methylation counts and
returns PIM score for these.
The function is used in calculateHistRPIM function.
}
