% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepIM}
\alias{prepIM}
\title{Given bisulfite sequencing data, prepares the intermediate methylation (IM) table}
\usage{
prepIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"))
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}
}
\description{
Given bisulfite sequencing data, prepares the intermediate methylation (IM) table
}
