% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepNorm.R
\name{prepNormIM}
\alias{prepNormIM}
\title{Prepares the normalized intermediate methylation (normIM) table}
\usage{
prepNormIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"),
  confLevel = 0.95, sdNorm = 0.13, samplingRate = 100)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{cacheDir}{If using caching, this argument specifies the directory to
use for storing the cache;
defaults to global option for \code{RESOURCES.RACHE},
if no such option has been specified you must provide one}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}

\item{sdNorm}{A decimal indicating the standard deviation of the normal 
distribution density function used for score weighting (the CpGs 
with 50% methylation get always highest weight indicationg highest 
heterogeneity, a low sdNorm gives a steeper weighting function which
gives only samples close to 50% methylation level high score); 
default is 0.13}

\item{samplingRate}{An integer deciding how finely are the generated 
density distributions sampled (how many points do we get in the 
interval 0-1 for each distribution); default is 100}
}
\value{
A \code{data.table} object with the following columns:
\itemize{
 \item{chr} {chromosome of methylation read}
 \item{start} {starting position for methylation read}
 \item{score} {normalized intermediate methylation score}
 }
}
\description{
Prepares the normalized intermediate methylation (normIM) table
}
