% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{promoter}
\alias{promoter}
\title{promoter}
\format{A large GRanges with 63344 elements:
\describe{
  \item{seqnames}{seqnames}
  \item{ranges}{ranges}
  \item{strand}{strand}
  \item{elementMetadata}{elementMetadata}
  \item{seqinfo}{seqinfo}
  \item{metadata}{metadata}
}}
\usage{
data(promoter)
promoter
}
\value{
A large GRanges object
}
\description{
promoter region annotation file
}
\keyword{datasets}
