% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGR.R
\name{readGR}
\alias{readGR}
\title{Make GenomicRanges Object}
\usage{
readGR(files, ids, n)
}
\arguments{
\item{files}{A vector of files containing methcolone output}

\item{ids}{A vector of sample ids for the files}

\item{n}{The index of the file vector to be read}
}
\value{
A GenomicRanges object containing pdr, epipolymorphism, and
Shannon entropy values for the nth file
}
\description{
Creates a GenomicRanges file for a singular methclone ouput file
}
\examples{
files <- c(system.file("extdata","D-2238.chr22.region.methClone_out.gz",package = "epihet"),
system.file("extdata","D-2668.chr22.region.methClone_out.gz",package = "epihet"),
system.file("extdata","N-1.chr22.region.methClone_out.gz",package = "epihet"),
system.file("extdata","N-2.chr22.region.methClone_out.gz",package = "epihet"))
ids <- epihet::splitn(basename(files),"[.]",1)
GR.Object <- epihet::readGR(files = files, ids = ids, n = 3)
}
