% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shannon.R
\name{shannon}
\alias{shannon}
\title{Shannon Entropy}
\usage{
shannon(p)
}
\arguments{
\item{p}{A vector of epiallele probabilities}
}
\value{
The Shannon entropy value
}
\description{
Calculates the Shannon entropy value
}
\examples{
a=c(rep(0,13),60.86960,0.00000,39.1304)
shannon(a)
}
