% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sharedmatrix}
\alias{sharedmatrix}
\title{sharedmatrix}
\format{A data frame with 479 rows and 8 variables:
\describe{
 \item{D-2238}{patients 1}
 \item{D-2668}{patients 2}
 \item{D-3314}{patients 3}
 \item{D-5360}{patients 4}
 \item{D-6947}{patients 5}
 \item{D-7076}{patients 6}
 \item{type}{type of epigenetic heterogeity}
 \item{location}{location of locus}
}}
\usage{
data(sharedmatrix)
sharedmatrix
}
\value{
A data frame
}
\description{
Epipolymorphism for 6 samples on DEH loci
}
\keyword{datasets}
