% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R, R/prepNorm.R
\name{singleSample}
\alias{singleSample}
\title{Check to see if bisulfite sequencing data has only one sample}
\usage{
singleSample(bsData)

singleSample(bsData)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{bsData}{Bisulfite sequencing data}
}
\value{
A boolean indicating whether or not the data is likely to only
represent a single sample

A boolean indicating whether or not the data is likely to only
represent a single sample
}
\description{
Check to see if bisulfite sequencing data has only one sample

Check to see if bisulfite sequencing data has only one sample
}
