% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitn.R
\name{splitn}
\alias{splitn}
\title{The subtring extraction of a character vector}
\usage{
splitn(strings, field, n)
}
\arguments{
\item{strings}{A GenomicRanges object to be compared}

\item{field}{A GenomicRanges object to be compared}

\item{n}{The value of gr1 to be compared}
}
\value{
A data frame containing a summary of the GenomicRanges object
}
\description{
Extract the subtrings of a character vector according to
the matches to substring split within them.
}
\examples{
x='chr1:10000-10005'
splitn(x,':',1)
}
