% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userobj.R
\name{userobj}
\alias{userobj}
\title{GenomicRanges object generation}
\usage{
userobj(data)
}
\arguments{
\item{data}{a data frame containing the chromosome number,
loci and strand information of DEH loci generated from diffHet() function.}
}
\value{
A GenomicRanges object
}
\description{
generate GenomicRanges object for DEH loci
}
\examples{
data=data.frame(chromosome=c('chr1','chr1','chr1'),
loci=c('6480531:6480554:6480561:6480565','6647655:6647696:6647701:6647705',
'7130155:7130172:7130179:7130188'),
strand=c('+','-','+'),stringsAsFactors = FALSE)
userobj(data)
}
