% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{BScredIntervalCache}
\alias{BScredIntervalCache}
\title{Calculate credibility interval based on binomial bayes distribution with caching}
\usage{
BScredIntervalCache(BSDT, cachedBinomialIntervals, hitCol = "methylCount",
  readCol = "coverage", confLevel = 0.95)
}
\arguments{
\item{BSDT}{Bisulfite sequencing data in a data.table format}

\item{cachedBinomialIntervals}{cachedBinomialIntervals}

\item{hitCol}{Name of column containing methylation count; defaults to "methylCount"}

\item{readCol}{Name of column containing coverage; defaults to "coverage"}

\item{confLevel}{The level of confidence to be used in the confidence interval; default is 0.95}
}
\description{
Calculate credibility interval based on binomial bayes distribution with caching
}
