% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{BScredIntervalCache}
\alias{BScredIntervalCache}
\title{Get credibility interval via binomial bayes distribution and caching}
\usage{
BScredIntervalCache(bsData, cachedBinomialIntervals,
  methylCol = "methylCount", coverageCol = "coverage",
  confLevel = 0.95)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{cachedBinomialIntervals}{cachedBinomialIntervals}

\item{methylCol}{Name of column containing methylation count;
defaults to "methylCount"}

\item{coverageCol}{Name of column containing coverage (i.e. number of reads);
defaults to "coverage"}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}
}
\value{
Bisulfite sequencing data as a \code{data.table} object with columns
indicating upper and lower limits of Bayesian binomial confidence interval
for methylation
}
\description{
Get credibility interval via binomial bayes distribution and caching
}
