% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R, R/prepNorm.R
\name{bsDataCheck}
\alias{bsDataCheck}
\title{Check bisulfite sequencing data and convert if needed}
\usage{
bsDataCheck(bsData)

bsDataCheck(bsData)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}

\item{bsData}{Bisulfite sequencing data}
}
\value{
A \code{list} of \code{data.table} objects that each contain bisulfite
sequencing data

A \code{list} of \code{data.table} objects that each contain bisulfite
sequencing data
}
\description{
Check bisulfite sequencing data and convert if needed

Check bisulfite sequencing data and convert if needed
}
