% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{bsDataCheck}
\alias{bsDataCheck}
\title{Helper function to check input Bisulfite sequencing data and convert as necessary}
\usage{
bsDataCheck(bsData)
}
\arguments{
\item{bsData}{Bisulfite sequencing data}
}
\description{
Helper function to check input Bisulfite sequencing data and convert as necessary
}
