% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{cacheBinomConfIntervals}
\alias{cacheBinomConfIntervals}
\title{Cache binomial confidence intervals}
\usage{
cacheBinomConfIntervals(maxHits, maxTotal, confLevel)
}
\arguments{
\item{maxHits}{maxHits}

\item{maxTotal}{maxTotal}

\item{confLevel}{The level of confidence to be used in the confidence interval; default is 0.95}
}
\description{
Cache binomial confidence intervals
}
