% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intervals.R
\name{cacheBinomConfIntervals}
\alias{cacheBinomConfIntervals}
\title{Cache binomial confidence intervals}
\usage{
cacheBinomConfIntervals(maxHits, maxTotal, confLevel)
}
\arguments{
\item{maxHits}{Maximum methylation count}

\item{maxTotal}{Maximum coverage}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}
}
\value{
A \code{data.table} object containing columns for upper and lower limits
of a Bayesian binomial confidence interval for maximum methylation count and
coverage; this serves as a cache that can replace the need to perform the
computationally expensive probability estimation
}
\description{
Cache binomial confidence intervals
}
