% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePIM.R
\name{calculatePIM}
\alias{calculatePIM}
\title{Given a Bisulfite data.table (BSDT), calculates the proportion
of intermediate methylation sites.}
\usage{
calculatePIM(BSDT, cache = TRUE)
}
\arguments{
\item{cache}{Logical indicating whether or not to use caching via \code{\link{simpleCache}}; default is TRUE}
}
\description{
TODO: Make the alpha level a parameter, both for confidence intervals,
and also for IM definition
}
