% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculatePIM}
\alias{calculatePIM}
\title{Given a Bisulfite data.table (BSDT), calculates the proportion
of intermediate methylation sites.}
\usage{
calculatePIM(BSDT, cache = TRUE, cacheDir = getOption("RESOURCES.RCACHE"))
}
\arguments{
\item{BSDT}{Bisulfite sequencing data in a data.table format}

\item{cache}{Logical indicating whether or not to use caching via \code{\link{simpleCache}}; default is TRUE}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}
}
\description{
Given a Bisulfite data.table (BSDT), calculates the proportion
of intermediate methylation sites.
}
