% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRPIM.R
\name{calculateRPIM}
\alias{calculateRPIM}
\title{Helper function to get the relative proportion of flagged sites for a single sample versus all other samples in a list of bisulfite data tables.}
\usage{
calculateRPIM(sampleName, BSDTsplit, cache = TRUE,
  cacheDir = getOption("RESOURCES.RCACHE"))
}
\arguments{
\item{sampleName}{The sample (which should specify a name in BSDTlist) to return
the proportion of sites for.}

\item{BSDTsplit}{A BSDT (bisulfite data.table) that has been split with
splitDataTable (so, a list of BSDTs); one corresponds to each sample to test.}

\item{cache}{Logical indicating whether or not to use caching via \code{\link{simpleCache}}; default is TRUE}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}
}
\description{
Helper function to get the relative proportion of flagged sites for a single sample versus all other samples in a list of bisulfite data tables.
}
