% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRPIM.R
\name{calculateRPIM}
\alias{calculateRPIM}
\title{Helper function to get the relative proportion of flagged sites for a single sample versus all other samples in a list of bisulfite data tables.}
\usage{
calculateRPIM(sampleName, BSDTsplit)
}
\arguments{
\item{sampleName}{The sample (which should specify a name in BSDTlist) to return
the proportion of sites for.}

\item{BSDTsplit}{A BSDT (bisulfite data.table) that has been split with
splitDataTable (so, a list of BSDTs); one corresponds to each sample to test.}
}
\description{
Helper function to get the relative proportion of flagged sites for a single sample versus all other samples in a list of bisulfite data tables.
}
