% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{background}
\alias{background}
\alias{datTraits}
\alias{DEG}
\alias{DEH}
\alias{diffhetmatrix}
\alias{moduledm}
\alias{modulesil}
\alias{promoter}
\alias{sharedmatrix}
\alias{myValues}
\title{example data
background}
\format{background: A data frame with 31995 rows and 1 variables:
\describe{
  \item{gene}{background gene list}
}}
\usage{
data(background)
background

data(datTraits)
datTraits

data(DEG)
DEG

data(DEH)
DEH

data(diffhetmatrix)
diffhetmatrix

data(moduledm)
moduledm

data(modulesil)
modulesil

data(promoter)
promoter

data(sharedmatrix)
sharedmatrix

data(myValues)
myValues
}
\value{
A data frame

A large GRanges object
}
\description{
background: A data frame containing 31995 elements as background used for pathway
enrichment analysis

datTraits: Clinical traits containing OS,EFS,age

DEG: Differentially expressed genes compared CEBPA-sil vs.normal

DEH: DEH loci

diffhetmatrix: A differentially heterogeneity matrix

moduledm: Module information for CEBPA-dm mutation samples

modulesil: Module information for CEBPA-sil mutation samples

promoter: The promoter region annotation file

sharedmatrix: Epigenetic heterogeneity values for 6 patients on DEH loci

myValues: customized epigenetic diversity metric for 4 samples
}
\keyword{datasets}
