% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRPIM.R
\name{getRPIM}
\alias{getRPIM}
\title{Get the relative proportion of flagged sites. This is a general version
of a method to get the RPIM (Relative Proportion of Intermediate Methylation).
Given a DT with bisulfite reads, and a flag column,
and then a huge table with this data for lots of samples, this will calculate
the relative proportion of CGs that are flagged, for pairwise comparisons
between all samples, subsetted to the CGs present in both samples.}
\usage{
getRPIM(BSDTsplit, cache = TRUE)
}
\arguments{
\item{BSDTsplit}{A BSDT (bisulfite data.table) that has been split with
splitDataTable (so, a list of BSDTs); one corresponds to each sample to test.}

\item{cache}{Logical indicating whether or not to use caching via \code{\link{simpleCache}}; default is TRUE}
}
\description{
Get the relative proportion of flagged sites. This is a general version
of a method to get the RPIM (Relative Proportion of Intermediate Methylation).
Given a DT with bisulfite reads, and a flag column,
and then a huge table with this data for lots of samples, this will calculate
the relative proportion of CGs that are flagged, for pairwise comparisons
between all samples, subsetted to the CGs present in both samples.
}
