% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGR.R
\name{makeGR}
\alias{makeGR}
\title{Make List of GenomicRanges Object}
\usage{
makeGR(files, ids, cores = 5, sve = FALSE)
}
\arguments{
\item{files}{A vector of input files containing bisulfite sequencing data}

\item{ids}{A vector of sample id's for the files}

\item{cores}{The number of cores to be used for parallel execution
(default: 5)}

\item{sve}{A boolean to save the GenomicRanges object (default: FALSE)}
}
\value{
A data frame of GenomicRanges objects containing pdr,
epipolymorphism, and Shannon entropy values for all input files.
Saves as an epi.gr.rda extension
}
\description{
Creates a GenomicRanges object for each bisulfite sequencing data
file
}
\examples{
path = system.file('extdata', package = 'epihet')
files = dir(path = path, pattern = '\\\\.bam$',
            recursive = TRUE, full.names = TRUE)[c(1,2)]
ids = basename(dirname(files))[c(1,2)]
GR.List = epihet::makeGR(files = files, ids = ids,
cores = 1, sve = FALSE)
}
