% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{PIM}
\alias{PIM}
\title{Given bisulfite sequencing data, calculates the proportion
of intermediate methylation sites.}
\usage{
PIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"), imLower = 0.25,
  imUpper = 0.75, confLevel = 0.95)
}
\arguments{
\item{bsData}{Bisulfite sequencing data;}

\item{cacheDir}{If using caching, this argument specifies the directory
to use for storing the cache;
defaults to global option for \code{RESOURCES.RACHE};
if no such option has been specified you must provide one}

\item{imLower}{The lower boundary for intermediate methylation (IM);
if a site is entirely below this threshold
(or if any part of its binomial credibility interval overlaps this boundary)
it is not considered IM;
defaults to .25}

\item{imUpper}{The upper boundary for intermediate methylation (IM);
if a site is entirely above this threshold
(or if any part of its binomial credibility interval overlaps this boundary)
it is not considered IM;
defaults to .75}

\item{confLevel}{A decimal indicating the level of confidence
to be used while creating cached the binomial bayes credibility interval;
default is .95 for 95 percent confidence}
}
\value{
A single value (numeric vector of length 1) indicating the
proportion of intermediate methylation (PIM) of an individual sample
}
\description{
Given bisulfite sequencing data, calculates the proportion
of intermediate methylation sites.
}
\examples{

data("exampleBSDT", package = "RPIM")

if (require(simpleCache)) {
PIM(exampleBSDT, cacheDir = "~")

simpleCache::setSharedCacheDir("cache")
PIM(exampleBSDT)
PIM(exampleBSDT, cacheDir = "~", imLower = .2, imUpper = .8)
}
}
