% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prepIM}
\alias{prepIM}
\title{Given a Bisulfite data.table (BSDT), prepares the intermediate methylation (IM) table}
\usage{
prepIM(BSDT, cacheDir = getOption("RESOURCES.RCACHE"))
}
\arguments{
\item{BSDT}{Bisulfite sequencing data in a data.table format}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}
}
\description{
Given a Bisulfite data.table (BSDT), prepares the intermediate methylation (IM) table
}
