% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepIM.R
\name{prepIM}
\alias{prepIM}
\title{Given a Bisulfite data.table (BSDT), prepares the intermediate methylation (IM) table}
\usage{
prepIM(BSDT)
}
\description{
Given a Bisulfite data.table (BSDT), prepares the intermediate methylation (IM) table
}
