% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGR.R
\name{readGR}
\alias{readGR}
\title{Make GenomicRanges Object}
\usage{
readGR(files, ids, n)
}
\arguments{
\item{files}{A vector of files containing bisulfite sequencing data}

\item{ids}{A vector of sample id's for the files}

\item{n}{The index of the file vector to be read}
}
\value{
A GenomicRanges object containing pdr, epipolymorphism, and
Shannon entropy values for the nth file
}
\description{
Creates a GenomicRanges file for a singular bisulfite sequencing
data file
}
\examples{
path = system.file('extdata', package = 'epihet')
files = dir(path = path, pattern = '\\\\.bam$',
            recursive = TRUE, full.names = TRUE)
ids = basename(dirname(files))
GR.Object = epihet::readGR(files = files, ids = ids, n = 3)
}
