% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{RPIM}
\alias{RPIM}
\alias{RPIM-package}
\title{Relative Proportion of sites with Intermediate Methylation (RPIM)}
\description{
RPIM is a score that measures the epigenetic heterogeneity in a
bisulfite sequencing sample. Under the assumption that a homogeneous
sample will have mostly CpGs with either 100% or 0% DNA methylation,
it follows that the proportion of sites that differ from these
two extremes can be used as a measure of sample heterogeneity.
}
\details{
This script (an incipient R package) provides functions
for assessing the RPIM given input DNA methylation calls.
}
\references{
\url{http://github.com/databio/RPIM}
}
