% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R, R/package.R
\docType{package}
\name{RPIM}
\alias{RPIM}
\alias{RPIM}
\alias{RPIM-package}
\title{Get the relative proportion of flagged sites.}
\usage{
RPIM(bsData, cacheDir = getOption("RESOURCES.RCACHE"), imLower = 0.25,
  imUpper = 0.75)
}
\arguments{
\item{bsData}{Bisulfite sequencing data for multiple samples; a BSDT (bisulfite data.table) that has been split with splitDataTable (so, a list of BSDTs); one corresponds to each sample to test.}

\item{cacheDir}{If using caching, this argument specifies the directory to use for storing the cache; defaults to global option for \code{RESOURCES.RACHE}, if no such option has been specified you must provide one}

\item{imLower}{The lower boundary for intermediate methylation (IM); if a site is entirely below this threshold (or if any part of a its binomial credibilty interval overlaps this boundary) it is not considered IM; defaults to .25}

\item{imUpper}{The upper boundary for intermediate methylation (IM); if a site is entirely above this threshold (or if any part of a its binomial credibilty interval overlaps this boundary) it is not considered IM; defaults to .75}
}
\description{
Get the relative proportion of flagged sites.

RPIM is a score that measures the epigenetic heterogeneity in a
bisulfite sequencing sample. Under the assumption that a homogeneous
sample will have mostly CpGs with either 100% or 0% DNA methylation,
it follows that the proportion of sites that differ from these
two extremes can be used as a measure of sample heterogeneity.
}
\details{
This script (an incipient R package) provides functions
for assessing the RPIM given input DNA methylation calls.
}
\references{
\url{http://github.com/databio/RPIM}
}
